services/ingestion-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-pubsub==2.27.2
pydantic==2.10.4
